console ui: improve the auto color detection, part 1/2
The current logic fails to detect color support in a number of
environments (notably it fails with rxvt families and tmux).

As almost all terminals support (at least some) coloring nowadays, it's
fairly reasonable to enable color output unless it is a 'dumb' terminal.
(in fact, this is exactly how many programs handle the $TERM variable).

This patch applies the change above to ConsoleTestRunner.