console ui: improve the auto color detection, part 2/2
This patch disables color escapes when the output does not go to a tty,
which makes these trivial tasks just work fine:

    $ ./run-test.py -v v | less  # pipe to another program
    $ ./run-test.py > result     # redirect to a file

Note: This behavior (suppressing auto-coloring if it is not connected
to a tty) is fairly common among unix commands. For example, `ls` and
`grep` with --color=auto option behave exactly like this, so this is
likely to be what most users expect.