Implement a context manager for exception-related testing.
This patch adds a new assertion method `assert_raise()` which can be
used as a context manager to test if a specific exception is raised
as expected.

For example, we can check whether some transactional commit raises an
error as below:

    def test_invalid_type(self):
        with self.assert_raise(InvalidTypeError):
            item = self.table.new_item()
            item['age'] = 'invalid'
            item.save()

This feature would be convenient for some use cases, as this notation
doesn't require users to define a test function.